Move tax_unit_is_joint variable to own module
